+++ sdccc/src/test/java/com/draeger/medical/sdccc/messages/TestMessageStorage.java
@@ -9,15 +9,18 @@ package com.draeger.medical.sdccc.messages;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTimeoutPreemptively;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
-import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
@@ -104,6 +107,9 @@ public class TestMessageStorage {
                     + "</msg:MetricState>"
                     + "</msg:ReportPart>"
                     + "</msg:EpisodicMetricReport>";
+    private static final byte[] ACTION_ENVELOPE = String.format(
+                    BASE_MESSAGE_STRING, "action", String.format(SEQUENCE_ID_METRIC_BODY_STRING, "3", "1"))
+            .getBytes(StandardCharsets.UTF_8);
     private static final String SEQUENCE_ID_ALERT_BODY_STRING =
             "<msg:EpisodicAlertReport MdibVersion=\"%s\" SequenceId=\"urn:uuid:%s\">"
                     + "<msg:ReportPart>"
@@ -257,9 +263,7 @@ public class TestMessageStorage {
                     CommunicationLog.MessageType.REQUEST,
                     headerContext,
                     messageStorage)) {
-                message.write(String.format(
-                                BASE_MESSAGE_STRING, "action", String.format(SEQUENCE_ID_METRIC_BODY_STRING, "3", "1"))
-                        .getBytes(StandardCharsets.UTF_8));
+                message.write(ACTION_ENVELOPE);
             }
             messageStorage.flush();
 
@@ -1847,6 +1851,7 @@ public class TestMessageStorage {
 
     /**
      * Tests if determineCharsetFromMessage() correctly determines the Charset when it is given in the HTTP Header.
+     *
      * @param dir - temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -1875,8 +1880,7 @@ public class TestMessageStorage {
 
             // then
             assertEquals(Charset.forName("ISO-8859-13"), actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
@@ -1923,6 +1927,7 @@ public class TestMessageStorage {
 
     /**
      * Ensures that no charset determination is performed if the EnableMessageEncodingCheck configuration is set to false.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -1999,13 +2004,14 @@ public class TestMessageStorage {
 
             // then
             assertEquals(StandardCharsets.UTF_8, actualCharset);
-            Mockito.verifyNoInteractions(this.testRunObserver);
+            verifyNoInteractions(this.testRunObserver);
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the HTTP Header and
      * Quotes are used.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2033,14 +2039,14 @@ public class TestMessageStorage {
             final Charset actualCharset = messageStorage.determineCharsetFromMessage(message);
             // then
             assertEquals(Charset.forName("ISO-8859-13"), actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the HTTP Header and
      * Double-Quotes are used.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2070,14 +2076,14 @@ public class TestMessageStorage {
 
             // then
             assertEquals(Charset.forName("ISO-8859-13"), actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the HTTP Header and
      * a boundary is given as well.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2106,14 +2112,14 @@ public class TestMessageStorage {
 
             // then
             assertEquals(Charset.forName("ISO-8859-13"), actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the HTTP Header and
      * a boundary is given as well.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2142,13 +2148,13 @@ public class TestMessageStorage {
 
             // then
             assertEquals(Charset.forName("ISO-8859-13"), actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the XML Declaration.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2180,14 +2186,14 @@ public class TestMessageStorage {
 
             // then
             assertEquals(StandardCharsets.ISO_8859_1, actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects an EBCDIC Charset when it is given in the
      * XML Declaration.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2219,14 +2225,14 @@ public class TestMessageStorage {
 
             // then
             assertEquals(Charset.forName("ebcdic-gb-285+euro"), actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the XML Declaration
      * using single quotes.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2259,13 +2265,13 @@ public class TestMessageStorage {
 
             // then
             assertEquals(StandardCharsets.ISO_8859_1, actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Tests if determineCharsetFromMessage() correctly detects the Charset when it is given in the Byte Order Mark.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2311,10 +2317,10 @@ public class TestMessageStorage {
             // then
             assertEquals(charset, actualCharset);
             if (expectFailure) {
-                Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
+                verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
                         .invalidateTestRun(anyString());
             } else {
-                Mockito.verifyNoInteractions(this.testRunObserver);
+                verifyNoInteractions(this.testRunObserver);
             }
         }
     }
@@ -2322,6 +2328,7 @@ public class TestMessageStorage {
     /**
      * Tests if determineCharsetFromMessage() invalidates the TestRun when the Charset of a message cannot be
      * determined.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2355,14 +2362,14 @@ public class TestMessageStorage {
 
             // then
             assertEquals(StandardCharsets.UTF_8, actualCharset);
-            Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                    .invalidateTestRun(anyString());
+            verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
         }
     }
 
     /**
      * Ensures that determineCharsetFromMessage() does not invalidate the TestRun when there are multiple Charset
      * declarations within a message that are consistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2377,12 +2384,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(StandardCharsets.UTF_8, actualCharset);
-        Mockito.verifyNoInteractions(this.testRunObserver);
+        verifyNoInteractions(this.testRunObserver);
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2398,13 +2406,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(Charset.forName("CP1147"), actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2420,13 +2428,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(StandardCharsets.ISO_8859_1, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2442,13 +2450,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(StandardCharsets.ISO_8859_1, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2463,13 +2471,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(StandardCharsets.UTF_8, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2485,13 +2493,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(StandardCharsets.UTF_8, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2507,13 +2515,13 @@ public class TestMessageStorage {
                 "application/soap+xml");
         // then
         assertEquals(StandardCharsets.UTF_8, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when the given Mime Type is not
      * standard-compliant.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2528,13 +2536,13 @@ public class TestMessageStorage {
                 "text/xml");
         // then
         assertEquals(StandardCharsets.UTF_8, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
     /**
      * Checks that determineCharsetFromMessage() invalidate the TestRun when there are multiple Charset
      * declarations within a message that are inconsistent.
+     *
      * @param dir - a temporary directory.
      * @throws IOException - when something goes wrong.
      */
@@ -2544,10 +2552,333 @@ public class TestMessageStorage {
                 dir, null, null, StandardCharsets.ISO_8859_1, StandardCharsets.UTF_8, "application/soap+xml");
         // then
         assertEquals(StandardCharsets.ISO_8859_1, actualCharset);
-        Mockito.verify(this.testRunObserver, VerificationModeFactory.atLeastOnce())
-                .invalidateTestRun(anyString());
+        verify(this.testRunObserver, VerificationModeFactory.atLeastOnce()).invalidateTestRun(anyString());
     }
 
+    /**
+     * Checks that convertMessageToMessageContent() works when it does not detect any decoding problems in the message.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertMessageToMessageContent(@TempDir final File dir) throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.UTF_8;
+        final Charset charsetInXMLDeclaration = StandardCharsets.UTF_8;
+        final Charset charsetInXMLDeclarationEncoding = StandardCharsets.UTF_8;
+        final String mimeType = "application/xml";
+
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, false, true, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag></sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(charsetInXMLDeclarationEncoding);
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+
+            // then
+            assertNotNull(result);
+            assertEquals(content, result.getBody());
+            verifyNoInteractions(testRunObserver);
+        }
+    }
+
+    /**
+     * Checks that convertMessageToMessageContent() uses the detected encoding.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertMessageToMessageContentUseAnotherEncoding(@TempDir final File dir) throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.UTF_16LE;
+        final Charset charsetInXMLDeclaration = StandardCharsets.UTF_16LE;
+        final Charset charsetInXMLDeclarationEncoding = StandardCharsets.UTF_16LE;
+        final String mimeType = "application/xml";
+
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, false, true, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag></sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(charsetInXMLDeclarationEncoding);
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+
+            // then
+            assertNotNull(result);
+            assertNotEquals(new String(encodedContent, StandardCharsets.UTF_8), result.getBody());
+            assertEquals(content, result.getBody());
+        }
+    }
+
+    /**
+     * Checks that convertMessageToMessageContent() fails when it detects a decoding problem in the message.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertToMessageContentFailOnEncodingProblem(@TempDir final File dir) throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.US_ASCII;
+        final Charset charsetInXMLDeclaration = StandardCharsets.US_ASCII;
+        final String mimeType = "application/xml";
+
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, false, true, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag>üöä</sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(StandardCharsets.UTF_8);
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+            // then
+            assertNotNull(result);
+        }
+        verify(testRunObserver, times(4)).invalidateTestRun(anyString());
+    }
+
+    /**
+     * Checks that convertMessageToMessageContent() fails when it detects a decoding problem in the message.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertToMessageContentFailOnEncodingProblem2(@TempDir final File dir) throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.UTF_8;
+        final Charset charsetInXMLDeclaration = StandardCharsets.UTF_8;
+        final String mimeType = "application/xml";
+
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, false, true, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag>üöä</sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(StandardCharsets.UTF_16);
+
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+
+            // then
+            assertNotNull(result);
+        }
+        verify(testRunObserver, times(6)).invalidateTestRun(anyString());
+    }
+
+    /**
+     * Checks that convertMessageToMessageContent() fails when it detects a decoding problem in the message.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertToMessageContentFailOnEncodingProblem3(@TempDir final File dir) throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.UTF_8;
+        final Charset charsetInXMLDeclaration = StandardCharsets.UTF_8;
+        final String mimeType = "application/xml";
+
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, false, true, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag>üöä</sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(StandardCharsets.ISO_8859_1);
+
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+
+            // then
+            assertNotNull(result);
+        }
+        verify(testRunObserver, times(2)).invalidateTestRun(anyString());
+    }
+
+    /**
+     * Checks that convertMessageToMessageContent() does not fail the TestRun, but correctly counts the encoding errors
+     * when SummarizeEncodingErrors is enabled.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertToMessageContentWhenSummarizeEncodingProblemsIsEnabled(@TempDir final File dir)
+            throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.UTF_8;
+        final Charset charsetInXMLDeclaration = StandardCharsets.UTF_8;
+        final String mimeType = "application/xml";
+
+        final MessageStorage storage;
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, true, true, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            storage = messageStorage;
+            assertEquals(0, messageStorage.getMessageEncodingErrorCount());
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag>üöä</sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(StandardCharsets.ISO_8859_1);
+
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+
+            // then
+            assertNotNull(result);
+        }
+        assertEquals(2, storage.getMessageEncodingErrorCount());
+    }
+
+    /**
+     * Checks that convertMessageToMessageContent() does not fail when the Encodingcheck is disabled.
+     *
+     * @param dir - a temporary directory.
+     * @throws IOException - when something goes wrong.
+     */
+    @Test
+    public void testConvertToMessageContentDoNotFailWhenEncodingCheckIsDisabled(@TempDir final File dir)
+            throws IOException {
+
+        final Charset charsetInHttpHeader = StandardCharsets.UTF_8;
+        final Charset charsetInXMLDeclaration = StandardCharsets.UTF_8;
+        final String mimeType = "application/xml";
+
+        try (final MessageStorage messageStorage = new MessageStorage(
+                1, false, false, mock(MessageFactory.class), new HibernateConfigImpl(dir), this.testRunObserver)) {
+            // given
+            final ListMultimap<String, String> headers = ArrayListMultimap.create();
+            headers.put("Content-Type", String.format("%s;charset=%s", mimeType, charsetInHttpHeader));
+
+            final HttpApplicationInfo applicationInfo = new HttpApplicationInfo(headers, "transactionId", "requestURI");
+            final TransportInfo transportInfo =
+                    new TransportInfo("http", "localhost", 1234, "remotehost", 4567, List.of());
+            final CommunicationContext communicationContext = new CommunicationContext(applicationInfo, transportInfo);
+            final Message message = new Message(
+                    CommunicationLog.Direction.INBOUND,
+                    CommunicationLog.MessageType.REQUEST,
+                    communicationContext,
+                    messageStorage);
+            final String content = String.format(
+                    "<?xml version=\"1.0\" encoding=\"%s\"?>%n<sometag>üöä</sometag>", charsetInXMLDeclaration);
+            final byte[] encodedContent = content.getBytes(StandardCharsets.ISO_8859_1);
+
+            message.write(encodedContent, 0, encodedContent.length);
+            message.close();
+
+            // when
+            final var result = messageStorage.convertMessageToMessageContent(message);
+
+            // then
+            assertNotNull(result);
+        }
+        verifyNoInteractions(testRunObserver);
+    }
+
+    /**
+     * Checks that determineCharsetFromMessage() works with the given Charsets.
+     *
+     * @param dir                             - a temporary directory.
+     * @param charsetInHttpHeader             - the Charset contained in the HTTP Header.
+     * @param bom                             - the ByteOrderMark in the message body.
+     * @param charsetInXMLDeclaration         - the Charset declared in the XML Declaration.
+     * @param charsetInXMLDeclarationEncoding - the Charset in which the XML Declaration is encoded.
+     * @throws IOException - when something goes wrong.
+     */
     private Charset testDetermineCharsetFromMessageUsingCharsets(
             final File dir,
             @Nullable final Charset charsetInHttpHeader,